refactor: Move defaultGuessPaths to favicons/defaults
