+++ src/exports/favicons.ts
@@ -4,6 +4,6 @@ export {
   type FaviconApiProvider,
   googleS2,
 } from '../favicons/api/index.js'
-export { defaultIconRels } from '../favicons/defaults.js'
+export { defaultGuessPaths, defaultIconRels } from '../favicons/defaults.js'
 export * from '../favicons/discover/types.js'
-export { defaultGuessPaths, type GuessMethodOptions } from '../favicons/guess/index.js'
+export type { GuessMethodOptions } from '../favicons/guess/index.js'
